refactor: colocate historical schemas with conversions
Refs #373
