+++ infrastructure/storage/document-conversions/006-learning-document.ts
@@ -1,7 +1,7 @@
 import type { z } from 'zod';
 import { learningDocumentSchema as outputSchema } from '@/domain/learning-document';
+import { outputSchema as inputSchema } from './005-combine-gist-connection';
 import { assertSchema } from './schema-utils';
-import { datasetV5Schema as inputSchema } from './schemas';
 
 export { inputSchema, outputSchema };
 export type Input = z.infer<typeof inputSchema>;
